swf: Extract `Twips` to a separate file
